# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for RaiseChild Server",
        "title": "RaiseChild Server API",
        "contact": {},
        "version": "1.0"
    },
    "host": "locahost:8080",
    "basePath": "/",
    "paths": {
...
        "/tx/build/money": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Prepares and builds a transaction for money actions (e.g., donate, withdraw) on-chain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Build a money-related transaction",
                "parameters": [
                    {
                        "description": "Money action details (e.g., amount, sender)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.MoneyActionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BuildTransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "401": {
                        "description": "You have no rights to access this action.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "500": {
                        "description": "There is something wrong in the system during the process. Please try again later.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    }
                }
            }
        },
        "/tx/execute": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Processes a transaction request and executes it on the Sui Network blockchain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Execute an on-chain transaction",
                "parameters": [
                    {
                        "description": "Transaction execution details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ExecuteTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "401": {
                        "description": "You have no rights to access this action.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    },
                    "500": {
                        "description": "There is something wrong in the system during the process. Please try again later.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageAPIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "request.ExecuteTransactionRequest": {
            "type": "object",
            "required": [
                "signature",
                "tx_bytes"
            ],
            "properties": {
                "signature": {
                    "type": "string"
                },
                "tx_bytes": {
                    "type": "string"
                }
            }
        },
...
        "request.MoneyActionRequest": {
            "type": "object",
            "required": [
                "action_type",
                "amoun",
                "sender"
            ],
            "properties": {
                "action_type": {
                    "description": "\"donate\" || \"withdraw\"",
                    "type": "string"
                },
                "amoun": {
                    "type": "integer",
                    "minimum": 1
                },
                "coin_type": {
                    "description": "e.g. \"SUI\", \"USDC\"",
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "sender": {
                    "type": "string"
                }
            }
        },
...
        "response.BuildTransactionResponse": {
            "type": "object",
            "properties": {
                "tx_bytes": {
                    "type": "string"
                }
            }
        },
...
        "response.MessageAPIResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}